use exec.Command on all platforms
