+++ internal/mongosh/mongosh.go
@@ -15,6 +15,7 @@
 package mongosh
 
 import (
+	"os"
 	"os/exec"
 )
 
@@ -28,6 +29,15 @@ func binPath() string {
 	}
 
 	return ""
+}
+
+func execCommand(args ...string) error {
+	cmd := exec.Command(mongoshBin, args...)
+	cmd.Stdin = os.Stdin
+	cmd.Stdout = os.Stdout
+	cmd.Stderr = os.Stderr
+
+	return cmd.Run()
 }
 
 func SetTelemetry(enable bool) error {
